+++ doc/release-notes.md
@@ -1,5 +1,5 @@
 # 1.0.1
-- upgrade to 1.0.1
+- refactoring and unit tests
 
 # 1.0.0
 - initial version