+++ doc/release-notes.md
@@ -1,3 +1,6 @@
+# 1.0.2
+- anonymous files handling enabled
+
 # 1.0.1
 - refactoring and unit tests